refactor: 연관성 높은 객체 별로 패키지 분리
