+++ src/main/java/lotto/model/lotto/Lottos.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.lotto;
 
 import java.util.ArrayList;
 import java.util.List;
