+++ src/main/java/lotto/model/winning/WinningResultResponses.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.winning;
 
 import java.util.List;
