+++ src/test/java/lotto/model/winning/RankTest.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.winning;
 
 import static org.assertj.core.api.Assertions.assertThat;
